refactor(tools): promotion, promote-entry and promotion-shared take shared-namespace names from their deps instead of src (L5, #864)
The three promotion modules imported shared-namespace resolution from
`src/shared-namespace.ts`, which reads its own environment. The server twin
`server/tools/shared-namespace.ts` takes the validated name set as a trailing
argument and throws when a call site omits it, so a module that forgets the
wiring fails by name rather than resolving against a second parse of the
environment.

All three now import the twin and pass `dependencies.sharedNamespaceNames`,
which `server/main.ts` already composes into `MemoryToolDependencies`:

- `promotion.ts` threads the name set through `foldNamespaceCounts` (for
  `canonicalNamespace`) and `authorizePromotion`.
- `promote-entry.ts` threads it through `resolvePromotionTarget`.
- `promotion-shared.ts` only changes the source of its `SharedNamespaceConfig`
  type; it already received the resolved config as a parameter, and its two
  callers now build that object from the deps-supplied names.

No default, no fallback, and no environment read is introduced. The resolved
names are the same values in an unconfigured deployment, so behavior is
unchanged; only their source moves to the composition root.

Two test files register these tools and so must supply the set the same way
`server/main.ts` does, using the existing
`server/tools/shared-namespace-fixture.ts`:

- `promotion.pg.test.ts`, named in the work item.
- `ported-tools-scope.test.ts`, which the full suite caught: it exercises the
  `promote_entry` legacy-target refusal, so the twin's missing-names error
  surfaced there.

`server/contracts/registered-tools.ts` needed no change: it stores handlers
without invoking them, so the names are never dereferenced.

Verified: bunx tsc --noEmit clean; oxlint --deny-warnings clean on the three
modules; bun run test:isolated 3944 pass / 0 fail / 35 skip across 261 files.

Refs #864
